LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/auto-009-expected.txt


FAIL Only recording last remembered inline size assert_equals: Using last remembered inline size - clientWidth expected 100 but got 0
FAIL Only recording last remembered block size assert_equals: Using last remembered block size - clientWidth expected 2 but got 0
FAIL contain:inline-size prevents recording last remembered inline size assert_equals: Size containment for inline axis - clientWidth expected 20 but got 0
FAIL contain:inline-size can keep previous last remembered inline size assert_equals: Size containment for inline axis - clientWidth expected 20 but got 0
